Move SettingsRow inline styles to CSS Module; split session into provider + hooks
SettingsRow.tsx was the last file using non-trivial inline-style layout —
moved to SettingsRow.module.css for consistency with the rest of the
codebase. While there, also corrected the import sort autofix on a couple
of files touched in this branch.

session.tsx tripped react-refresh/only-export-components by exporting both
SessionProvider (a component) and the four session hooks. Split into:
  src/app/session.ts           — SessionContext + hooks (no JSX)
  src/app/SessionProvider.tsx  — provider component only
Functionally identical; satisfies --max-warnings 0 lint gate.
