+++ src/features/board-builder/SettingsRow.module.css
@@ -1,0 +1,18 @@
+.row {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+  margin-top: 12px;
+  margin-bottom: 28px;
+}
+
+.spacer {
+  flex: 1;
+}
+
+.count {
+  align-self: center;
+  font-size: 13px;
+  color: var(--tal-ink-muted);
+  font-weight: 600;
+}
